% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_setup.R
\docType{data}
\name{pkg.env}
\alias{pkg.env}
\title{Initial setup to terrahidro package}
\format{
An object of class \code{environment} of length 0.
}
\usage{
pkg.env
}
\description{
Initial setup to terrahidro package
}
\keyword{datasets}
