% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_carvev.R
\name{th_carvev}
\alias{th_carvev}
\title{Carve channels on a MDT in a v-shaped form}
\usage{
th_carvev(dem, output, keep_mdt = TRUE)
}
\arguments{
\item{dem}{May be a SpatRaster object or a file path indicating the source of the MDT}

\item{output}{Specify the output of the carved MDT}

\item{keep_mdt}{Keep the original MDT in temp directory for later use}
}
\value{
A SpatRaster object of the v-shaped carved MDT
}
\description{
Carve channels on a MDT in a v-shaped form
}
\examples{
# carved <- th_carvev(srtm1arc, "~/dem_carvev.tif")
}
