% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_workflow}
\alias{update_workflow}
\title{Update workflow}
\usage{
update_workflow(layer, path)
}
\arguments{
\item{layer}{Layer in workflow}

\item{path}{Path of the layer in workflow}
}
\value{
TRUE if it was possible to update the workflow, FALSE otherwise.
}
\description{
Update workflow
}
